requirements.txt
aiohttp==3.8.5
aiosignal==1.3.1
async-timeout==4.0.3
attrs==23.1.0
certifi==2023.7.22
charset-normalizer==3.2.0
colorama==0.4.6
discord-py-interactions==5.10.0
discord-typings==0.7.0
emoji==2.8.0
frozenlist==1.4.0
idna==3.4
multidict==6.0.4
openai==0.28.0
python-dotenv==1.0.0
requests==2.31.0
tomli==2.0.1
tqdm==4.66.1
typing_extensions==4.7.1
urllib3==2.0.4
yarl==1.9.2
pytest
